Server to cater to Lambda Proxy response
If lambda proxy is enabled (which is by default in Serverless),
the response returned in the lambda callback must be of the form:

```json
{
  statusCode: 200,
  body: { /** data **/ }
}
```

The current server sends the entire response without separating
the statusCode and body which is what this commit does.

Ideally, we should detect if the lambda is using lambda proxy or not
but I cannot find the flag to detect this setting.