Issue #107 Hook on "deploy function" command
This fix hooks the webpack plugin onto the Serverless "deploy function" command. Currently the plugin will webpack the service as it would be done with a regular deploy. Then the single selected function is deployed by Serverless.
This makes the whole thing work. However there could be some improvements done later. Maybe _this.options.f_ can be used in this case to restrict the webpack run to create only the selected function entry point. But finding that would not be trivial.
I suggest to approve this solution on the first hand and handle the rest as improvement.