Make processed input path absolute (#156)
* Make processed input path absolute

When using the invoke local command with the --path (-p) flag, relative file paths will not be found because serverless will end up looking in the build directory later on. So, this commit ensures that the path is absolute, ensuring it will be found, even if the path was relative.

Couldn't think of a good name for the function. Also, I'm not sure that this should go in the utils file.

Happy to make revisions!

* Move execution building out of utils

* Add lib/makePathOptionAbsolute.js