Fix the `spawn E2BIG` error when zipping
When the native method to zip is used, we give it a list of all files to be included in the archive.
Because of `node_modules`, that list can be very huge and then we hit a limit which broke the build with `spawn E2BIG` error.

We list files because we might need to filter them if the `excludeRegex` option is used.
So, the zip method is now forced to node in case of that option is defined.
Otherwise, we are using the method available (which might be the native one in most case) and we don't give it a list of files to be zipped but a folder (we don't use `*` because it'll omit dot files).