Re-add the `--no-build` CLI option as `--skip-build`
We removed it because it was generating some deprecated message and we introduced the parameter `noBuild` instead.
Now the CLI option is back as a new name (see below) and the parameter is kept.

Serverless add an extra CLI option prefixed with `--no-` in case we add a new CLI option (ie: `--build`). But we don't need the `--build` option, only the `--no-build` which isn't possible. So we renamed it to `--skip-build`